refactor: modal context 분리
