+++ src/app/mission/[id]/stopwatch/Stopwatch.context.tsx
@@ -9,19 +9,14 @@ import {
   useEffect,
   useMemo,
 } from 'react';
-import { useRouter } from 'next/navigation';
-import { useRecordTime } from '@/apis';
-import { isSeverError } from '@/apis/instance.api';
-import { BackDialog, FinalDialog, MidOutDialog } from '@/app/mission/[id]/stopwatch/modals';
+import { useSubmit } from '@/app/mission/[id]/stopwatch/index.hooks';
+import ModalContextProvider from '@/app/mission/[id]/stopwatch/Modal.context';
 import Loading from '@/components/Loading';
 import { EVENT_LOG_CATEGORY, EVENT_LOG_NAME } from '@/constants/eventLog';
-import { ROUTER } from '@/constants/router';
 import useStopwatchSeconds from '@/hooks/mission/stopwatch/useStopwatchLogic';
 import useStopwatchStatus, { type StepType } from '@/hooks/mission/stopwatch/useStopwatchStatus';
-import useModal from '@/hooks/useModal';
 import { eventLogger } from '@/utils';
-import { getProgressMissionStartTimeToStorage, removeProgressMissionData } from '@/utils/storage/progressMission';
-import { formatDate, formatMMSS } from '@/utils/time';
+import { formatMMSS } from '@/utils/time';
 
 interface TimeContextProps {
   minutes: string;
@@ -47,18 +42,6 @@ export const StopwatchStepContext = createContext<StepContextProps>({
   step: 'ready',
   prevStep: 'ready',
   onNextStep: () => {},
-});
-
-interface ModalContextProps {
-  openMidOutModal: () => void;
-  openFinalModal: () => void;
-  openBackModal: () => void;
-}
-
-const ModalContext = createContext<ModalContextProps>({
-  openMidOutModal: () => {},
-  openFinalModal: () => {},
-  openBackModal: () => {},
 });
 
 function StopwatchProvider({
@@ -119,9 +102,7 @@ function StopwatchProvider({
     <>
       <StopwatchStepContext.Provider value={stepValue}>
         <StopwatchTimeContext.Provider value={timeValue}>
-          <ModalContextProvider missionId={missionId} onSubmit={onSubmit}>
-            {children}
-          </ModalContextProvider>
+          <ModalContextProvider missionId={missionId}>{children}</ModalContextProvider>
         </StopwatchTimeContext.Provider>
       </StopwatchStepContext.Provider>
       {isSubmitLoading && <Loading />}
@@ -143,134 +124,6 @@ export const useStopwatchStepContext = () => {
   const context = useContext(StopwatchStepContext);
   if (!context) {
     throw new Error('useStopwatchStepContext must be used within a StopwatchProvider');
-  }
-
-  return context;
-};
-
-const useSubmit = ({ missionId, second }: { missionId: string; second: number }) => {
-  const router = useRouter();
-
-  const { formattedMinutes, formattedSeconds } = formatMMSS(second);
-
-  const { mutate, isPending: isSubmitLoading } = useRecordTime({
-    onSuccess: (response) => {
-      const missionRecordId = String(response.missionId);
-      router.replace(ROUTER.RECORD.CREATE(missionRecordId));
-      eventLogger.logEvent('api/record-time', 'stopwatch', { missionRecordId });
-      removeProgressMissionData();
-    },
-    onError: (error) => {
-      if (isSeverError(error)) {
-        if (error.response.data.data.errorClassName === 'MISSION_RECORD_ALREADY_EXISTS_TODAY') {
-          removeProgressMissionData();
-          router.replace(ROUTER.HOME);
-        }
-      }
-    },
-  });
-
-  const onSubmit = async () => {
-    const startTimeString = getProgressMissionStartTimeToStorage(missionId);
-    if (!startTimeString) return;
-
-    const startTime = new Date(startTimeString);
-    const startTimeFormatted = formatDate(startTime);
-    const finishTimeFormatted = formatDate(new Date());
-
-    mutate({
-      missionId: missionId,
-      startedAt: startTimeFormatted,
-      finishedAt: finishTimeFormatted,
-      durationMin: Number(formattedMinutes),
-      durationSec: Number(formattedSeconds),
-    });
-  };
-
-  return {
-    isSubmitLoading,
-    onSubmit,
-  };
-};
-
-function ModalContextProvider({
-  children,
-  missionId,
-  onSubmit,
-}: PropsWithChildren<{
-  missionId: string;
-  onSubmit: () => void;
-}>) {
-  const router = useRouter();
-  const { time } = useStopwatchTimeContext();
-  const { prevStep, onNextStep } = useStopwatchStepContext();
-
-  const { isOpen: isFinalModalOpen, openModal: openFinalModal, closeModal: closeFinalModal } = useModal();
-  const { isOpen: isBackModalOpen, openModal: openBackModal, closeModal: closeBackModal } = useModal();
-  const { isOpen: isMidOutModalOpen, openModal: openMidOutModal, closeModal: closeMidOutModal } = useModal();
-
-  const value = useMemo(
-    () => ({
-      openMidOutModal,
-      openFinalModal,
-      openBackModal,
-    }),
-    [openBackModal, openFinalModal, openMidOutModal],
-  );
-
-  const logData = {
-    finishTime: time,
-  };
-
-  const onCancel = () => {
-    eventLogger.logEvent(EVENT_LOG_NAME.STOPWATCH.CLICK_CANCEL, EVENT_LOG_CATEGORY.STOPWATCH, logData);
-    onNextStep(prevStep);
-  };
-
-  const onFinish = () => {
-    // TODO: 끝내기 로직 추가
-    // 이쪽에 로딩 추가 필요
-    onSubmit();
-  };
-
-  // 뒤로가기 버튼 눌렀을 때
-  const onExit = () => {
-    router.replace(ROUTER.MISSION.DETAIL(missionId));
-    removeProgressMissionData();
-  };
-
-  return (
-    <ModalContext.Provider value={value}>
-      {children}
-      <FinalDialog
-        isOpen={isFinalModalOpen}
-        onClose={closeFinalModal}
-        onCancel={onCancel}
-        onConfirm={onFinish}
-        logData={logData}
-      />
-      <BackDialog
-        isOpen={isBackModalOpen}
-        onClose={closeBackModal}
-        onCancel={onCancel}
-        onConfirm={onExit}
-        logData={logData}
-      />
-      <MidOutDialog
-        isOpen={isMidOutModalOpen}
-        onClose={closeMidOutModal}
-        onCancel={onCancel}
-        onConfirm={onExit}
-        logData={logData}
-      />
-    </ModalContext.Provider>
-  );
-}
-
-export const useStopwatchModalContext = () => {
-  const context = useContext(ModalContext);
-  if (!context) {
-    throw new Error('useStopwatchModalContext must be used within a StopwatchProvider');
   }
 
   return context;
